+++ aa-display/src/main/java/io/github/nitsuya/aa/display/xposed/hook/aa/AaUiHook.kt
@@ -220,7 +220,6 @@ object AaUiHook: AaHook() {
         }
 
         val classes = bridge.findClass {
-            searchPackages = listOf("")
             matcher {
                 usingStrings {
                     add(